+++ web/style.ts
@@ -22,6 +22,7 @@ export default function highlightStyles(mdExtension: MDExt[]) {
     { tag: ct.AttributeNameTag, class: "sb-atom" },
     { tag: ct.TaskTag, class: "sb-task" },
     { tag: ct.TaskMarkerTag, class: "sb-task-marker" },
+    { tag: ct.TaskStateTag, class: "sb-task-state" },
     { tag: ct.CodeInfoTag, class: "sb-code-info" },
     { tag: ct.CommentTag, class: "sb-comment" },
     { tag: ct.CommentMarkerTag, class: "sb-comment-marker" },